+++ event_sourcery/async_/encryption.py
@@ -1,0 +1,9 @@
+__all__ = [
+    "AsyncEncryption",
+    "AsyncNoKeyStorageStrategy",
+]
+
+from event_sourcery._event_store._async.encryption import (
+    AsyncEncryption,
+    AsyncNoKeyStorageStrategy,
+)