+++ tests/event_store/event/test_privacy.py
@@ -6,10 +6,12 @@ import pytest
 from pydantic import BaseModel, Field
 
 from event_sourcery import Event, StreamId
+from event_sourcery.async_.backend import AsyncInMemoryKeyStorage
 from event_sourcery.backend import Backend, InMemoryKeyStorage
 from event_sourcery.encryption import DataSubject, Encrypted
 from event_sourcery.exceptions import KeyNotFoundError, NoSubjectIdFound
 from event_sourcery.interfaces import EncryptionStrategy
+from tests.adapter import BackendFacade
 from tests.bdd import Given, Then, When
 
 
@@ -291,8 +293,14 @@ def test_invalid_encryption_configuration(given: Given, when: When) -> None:
 
 @pytest.fixture()
 def backend(backend: Backend) -> Backend:
+    strategy = XorEncryptionStrategy()
+    if isinstance(backend, BackendFacade):
+        return backend.with_encryption(
+            strategy=strategy,
+            key_storage=AsyncInMemoryKeyStorage(),
+        )
     return backend.with_encryption(
-        strategy=XorEncryptionStrategy(),
+        strategy=strategy,
         key_storage=InMemoryKeyStorage(),
     )